feat: Access Token API

+++ src/pages/api/[[...user]].tsx
@@ -1,5 +1,6 @@
+import jwt from 'jsonwebtoken';
 import { deleteData, retrieveData, updateData } from '@/lib/firebase/service';
-import { NextApiRequest, NextApiResponse } from 'next';
+import type { NextApiRequest, NextApiResponse } from 'next';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   const users = await retrieveData('users');
@@ -16,36 +17,59 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
       data: users,
     });
   } else if (req.method === 'PUT') {
-    const { id, data } = req.body;
-    await updateData('users', id, data, (result: boolean) => {
-      if (result) {
-        res.status(200).json({
-          status: true,
-          statusCode: 200,
-          message: 'success',
+    const { user }: any = req.query;
+    const { data } = req.body;
+    const token = req.headers.authorization?.split(' ')[1] || '';
+    jwt.verify(token, process.env.NEXTAUTH_SECRET || '', async (err: any, decoded: any) => {
+      if (decoded && decoded.role === 'admin') {
+        await updateData('users', user[1], data, (result: boolean) => {
+          if (result) {
+            res.status(200).json({
+              status: true,
+              statusCode: 200,
+              message: 'success',
+            });
+          } else {
+            res.status(400).json({
+              status: false,
+              statusCode: 400,
+              message: 'failed',
+            });
+          }
         });
       } else {
-        res.status(400).json({
+        res.status(403).json({
           status: false,
-          statusCode: 400,
-          message: 'failed',
+          statusCode: 403,
+          message: 'Access Denied',
         });
       }
     });
   } else if (req.method === 'DELETE') {
     const { user }: any = req.query;
-    await deleteData('users', user[1], (result: boolean) => {
-      if (result) {
-        res.status(200).json({
-          status: true,
-          statusCode: 200,
-          message: 'success',
+    const token = req.headers.authorization?.split(' ')[1] || '';
+    jwt.verify(token, process.env.NEXTAUTH_SECRET || '', async (err: any, decoded: any) => {
+      if (decoded && decoded.role === 'admin') {
+        await deleteData('users', user[1], (result: boolean) => {
+          if (result) {
+            res.status(200).json({
+              status: true,
+              statusCode: 200,
+              message: 'success',
+            });
+          } else {
+            res.status(400).json({
+              status: false,
+              statusCode: 400,
+              message: 'failed',
+            });
+          }
         });
       } else {
-        res.status(400).json({
+        res.status(403).json({
           status: false,
-          statusCode: 400,
-          message: 'failed',
+          statusCode: 403,
+          message: 'Access Denied',
         });
       }
     });
